Split ErrorReportFormatter into title and body formatters
- Reduces indirection, parameter objects, and simplifies testing.
- Fix up ErrorReportBodyFormatterTest, consistently use mocks
