+++ game-core/src/main/java/org/triplea/debug/error/reporting/StackTraceReportModel.java
@@ -4,11 +4,13 @@ import games.strategy.triplea.ResourceLoader;
 import games.strategy.triplea.ui.UiContext;
 import java.util.Optional;
 import java.util.function.Consumer;
-import java.util.function.Function;
 import java.util.function.Predicate;
 import javax.annotation.Nonnull;
 import lombok.Builder;
 import org.triplea.debug.LoggerRecord;
+import org.triplea.debug.console.window.DebugUtils;
+import org.triplea.debug.error.reporting.formatting.ErrorReportBodyFormatter;
+import org.triplea.debug.error.reporting.formatting.ErrorReportTitleFormatter;
 import org.triplea.http.client.error.report.ErrorReportRequest;
 
 @Builder
@@ -16,7 +18,6 @@ class StackTraceReportModel {
 
   @Nonnull private final StackTraceReportView view;
   @Nonnull private final LoggerRecord stackTraceRecord;
-  @Nonnull private final Function<ErrorReportRequestParams, ErrorReportRequest> formatter;
   @Nonnull private final Predicate<ErrorReportRequest> uploader;
   @Nonnull private final Consumer<ErrorReportRequest> preview;
 
@@ -27,15 +28,17 @@ class StackTraceReportModel {
   }
 
   private ErrorReportRequest readErrorReportFromUi() {
-    return formatter.apply(
-        ErrorReportRequestParams.builder()
-            .userDescription(view.readUserDescription())
-            .mapName(
+    return ErrorReportRequest.builder()
+        .title(ErrorReportTitleFormatter.createTitle(stackTraceRecord))
+        .body(
+            ErrorReportBodyFormatter.buildBody(
+                view.readUserDescription(),
                 Optional.ofNullable(UiContext.getResourceLoader())
                     .map(ResourceLoader::getMapName)
-                    .orElse(null))
-            .logRecord(stackTraceRecord)
-            .build());
+                    .orElse(null),
+                DebugUtils.getMemory(),
+                stackTraceRecord))
+        .build();
   }
 
   void previewAction() {
